+++ SleepHunter/Services/Configuration/IHotkeyRegistrationService.cs
@@ -1,0 +1,11 @@
+﻿using SleepHunter.Macro;
+
+namespace SleepHunter.Services.Configuration
+{
+    public interface IHotkeyRegistrationService
+    {
+        bool Register(Hotkey hotkey);
+
+        bool Unregister(Hotkey hotkey);
+    }
+}